feat: 调整字段的类型

+++ src/main/java/com/ckb/explorer/entity/CkbTransaction.java
@@ -34,7 +34,7 @@ public class CkbTransaction {
 
   private byte[] headerDeps;
 
-  private Integer cycles;
+  private Long cycles;
 
   private Long transactionFee;
